Crate node_template_runtime

The Substrate Node Template runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

opaque
Opaque types. These are used by the CLI to instantiate machinery that don't need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

Macros

construct_runtime
Construct a runtime, with the given name and the given modules.

parameter_types
Macro for easily creating a new implementation of the Get trait. Use similarly to how you would declare a const:

Structs

BlockHashCount

CreationFee

ExistentialDeposit

GenesisConfig

MinimumPeriod

Perbill
Perbill is parts-per-billion. It stores a value between 0 and 1 in fixed point and provides a means to multiply some other value by that.

Permill
Permill is parts-per-million (i.e. after multiplying by this, divide by 1000000).

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

TransactionBaseFee

TransactionByteFee

TransferFee

Enums

BalancesCall

Call

Event

Origin

TimestampCall

Constants

VERSION
This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage
Complex storage builder stuff.

StorageValue
A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version
The version infromation used to identify this runtime when compiled natively.

Type Definitions

AccountId
Alias to pubkey that identifies an account on the chain.

AccountSignature
The type used by authorities to prove their ID.

Aura

AuraConfig

AuraId
The Ed25519 pub key of an session that belongs to an Aura authority of the chain.

AuraSignature
Alias to the signature scheme used for Aura authority signatures.

Balance
Balance type for the node.

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

BlockNumber
Index of a block number in the chain.

CheckedExtrinsic
Extrinsic type that has already been checked.

Executive
Executive: handles dispatch to the various modules.

Hash
A hash of some data used by the chain.

Header
Block header type as expected by this runtime.

Indices

IndicesConfig

Nonce
Index of an account's extrinsic in the chain.

Sudo

SudoConfig

System

SystemConfig

TemplateModule

Timestamp

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.