Module node_template_runtime::opaque

Opaque types. These are used by the CLI to instantiate machinery that don't need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

Structs

UncheckedExtrinsic
Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

Type Definitions

Block
Opaque block type.

BlockId
Opaque block identifier type.

Header
Opaque block header type.

SessionKey
Opaque session key type.